+++ packages/webcodecs/src/it-tests/upsample-audiodata.ts
@@ -2,8 +2,8 @@ import {resampleAudioData} from '../resample-audiodata';
 import {assertJson, audioDataToSerializable} from './assertions';
 
 const audioData = new AudioData({
-	data: new Float32Array([1, 2, 3]),
-	format: 'f32',
+	data: new Int32Array([1, 2, 3]),
+	format: 's32',
 	numberOfChannels: 1,
 	numberOfFrames: 3,
 	sampleRate: 22050,
@@ -14,7 +14,7 @@ assertJson(
 	audioDataToSerializable(resampleAudioData({audioData, newSampleRate: 44100})),
 	{
 		data: [[1, 1, 2, 2, 3, 3]],
-		format: 'f32',
+		format: 's32',
 		numberOfChannels: 1,
 		numberOfFrames: 6,
 		sampleRate: 44100,
